# public/swagger.yml
...
swagger: '2.0'
...
info:
  version: "1"
  title: Clowder API
  description: |
    Documentation for the [Clowder](http://clowder.ncsa.illinois.edu/) API

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
################################################################################
host: clowder.ncsa.illinois.edu
basePath: /clowder/api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################
...
securityDefinitions:
  login:
    type: basic
  key:
    type: apiKey
    name: key
    in: query
security:
  - key: []
  - login: []

################################################################################
# Parameters                                                                   #
################################################################################

################################################################################
# Paths                                                                        #
################################################################################
paths:
...
  /files/{file_id}/extractions:
    post:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: file_id
          in: path
      tags:
        - files
        - extractions
      summary: Submit file for extraction by a specific extractor
      description: Extractions for Files.
...
  /datasets/{ds_id}/extractions:
    post:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: ds_id
          in: path
      tags:
        - extractions
      summary: Submit dataset for extraction by a specific extractor
      description: Extractions for Files.
...